requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
idna==3.3
numpy==1.23.1
opencv-python==4.6.0.66
pika==1.3.0
Pillow==9.2.0
PyYAML==6.0
requests==2.28.1
urllib3==1.26.11
yadisk==1.2.15
